Move send_response_from_cache() into cli
