+++ crates/cli/src/command/grep/forerunner.rs
@@ -1,9 +1,9 @@
 use crate::app::Args;
+use crate::{send_response_from_cache, SendResponse};
 use anyhow::Result;
 use clap::Parser;
 use maple_core::process::CacheableCommand;
 use maple_core::tools::rg::{rg_command, rg_shell_command};
-use maple_core::utils::{send_response_from_cache, SendResponse};
 use std::path::PathBuf;
 use utils::is_git_repo;
