+++ crates/cli/src/lib.rs
@@ -3,3 +3,44 @@ pub mod command;
 
 /// Re-exports.
 pub use app::{Args, RunCmd};
+
+use icon::Icon;
+use std::path::Path;
+use utils::{println_json, println_json_with_length, read_first_lines};
+
+#[derive(Debug, Clone)]
+#[allow(unused)]
+pub enum SendResponse {
+    Json,
+    JsonWithContentLength,
+}
+
+/// Reads the first lines from cache file and send back the cached info.
+pub fn send_response_from_cache(
+    tempfile: &Path,
+    total: usize,
+    response_ty: SendResponse,
+    icon: Icon,
+) {
+    let using_cache = true;
+    if let Ok(iter) = read_first_lines(&tempfile, 100) {
+        let lines: Vec<String> = if let Some(icon_kind) = icon.icon_kind() {
+            iter.map(|x| icon_kind.add_icon_to_text(x)).collect()
+        } else {
+            iter.collect()
+        };
+        match response_ty {
+            SendResponse::Json => println_json!(total, tempfile, using_cache, lines),
+            SendResponse::JsonWithContentLength => {
+                println_json_with_length!(total, tempfile, using_cache, lines)
+            }
+        }
+    } else {
+        match response_ty {
+            SendResponse::Json => utils::println_json!(total, tempfile, using_cache),
+            SendResponse::JsonWithContentLength => {
+                println_json_with_length!(total, tempfile, using_cache)
+            }
+        }
+    }
+}
