+++ crates/maple_core/src/cache/mod.rs
@@ -1,6 +1,6 @@
 use crate::datastore::CACHE_INFO_IN_MEMORY;
 use crate::process::ShellCommand;
-use crate::utils::UtcTime;
+use crate::UtcTime;
 use chrono::prelude::*;
 use std::path::PathBuf;
